extract shared WidgetFrame from widget chrome
framing reused by the dashboard and preview hosts
