+++ packages/widget-dashboard/src/components/widget-chrome/widget-chrome.module.css
@@ -1,26 +1,11 @@
+/* contain: layout (not size) keeps a tall body inside the tile without breaking masonry measurement. */
 .widgetChrome {
 	height: 100%;
-
-	/*
-	 * Isolate the widget body from the surrounding layout. In a fixed-row
-	 * grid the tile height comes from its row span, but a body taller than
-	 * the tile lets its intrinsic size feed the grid item's `min-height: auto`
-	 * and overflow the cell, adding scroll space below the grid. Layout
-	 * containment severs that propagation so the body stays within the tile
-	 * and scrolls internally. It contains layout only, not size or paint, so
-	 * the masonry surface still measures the tile from its content and the
-	 * tile elevation is not clipped.
-	 */
 	contain: layout;
 }
 
-/*
- * When keyboard focus lands on the widget's own content (a link or control
- * the render module draws), outline the whole tile so Tab users can tell
- * which widget holds focus, since an inner ring alone is easy to miss. Scoped
- * to content focus, and to `:focus-visible` so a mouse click doesn't draw it.
- */
-.widgetChrome:has(.widgetChromeContent :focus-visible) {
+/* Outline the whole tile when the body holds keyboard focus; an inner ring is easy to miss. */
+.widgetChrome:has(:focus-visible) {
 	outline: var(--wpds-border-width-focus) solid var(--wpds-color-stroke-focus);
 	outline-offset: 2px;
 }
@@ -30,58 +15,14 @@
 	color: var(--wpds-color-foreground-content-neutral);
 }
 
-/*
- * Scroll container for the widget body. The card root clips its own overflow
- * and the header is a fixed-size flex sibling, so confining the scroll to this
- * element is what keeps the header pinned while the body scrolls.
- */
+/* Fill the tile so placeholder states center in the full height. */
 .widgetChromeContent {
 	flex: 1;
 	height: 100%;
-	overflow-y: auto;
-
-	/*
-	 * As a flex item, the content area defaults to `min-height: auto`, which
-	 * lets tall widget content expand it past the height the card allots
-	 * instead of staying bounded. Pinning the min sizes to 0 keeps the area
-	 * within its slot so a widget's own `overflow` scrolls internally rather
-	 * than pushing past the card.
-	 */
 	min-width: 0;
 	min-height: 0;
 }
 
-/*
- * Bleed block edges: zero the content's block padding so the body fills the
- * card top to bottom. The top is ours to flush (a full-bleed widget hides its
- * header in a position:absolute VisuallyHidden node that stays the card's
- * :first-child and blocks Card.FullBleed's top-flush rule). The bottom pairs
- * with the `margin-block-end: 0` on the bleed scroll wrapper below.
- */
-.widgetChromeContentBleed {
-	padding-block: 0;
-}
-
-/*
- * Scroll container for a bleeding body. Card.FullBleed only does the edge
- * breakout (negative margins) and declares no height or overflow, so the scroll
- * container must be re-established here, on the wrapper that spans the card
- * edges, so the scrollbar tracks the edge and the header stays pinned.
- *
- * Card.FullBleed also adds `margin-block-end: calc(var(--wp-ui-card-padding) *
- * -1)` to cancel the content's bottom padding; that padding is zeroed above, so
- * neutralize the margin. Left in, it pulls the wrapper's box short of the body,
- * surfacing as a phantom scrollbar on a content-driven tile or a bottom inset
- * on a fixed-height one.
- */
-.widgetChromeBleedScroll {
-	height: 100%;
-	overflow-y: auto;
-	margin-block-end: 0;
-}
-
-.loading,
 .unavailable {
 	height: 100%;
 }
-
