+++ packages/widget-dashboard/src/components/widget-chrome/widget-chrome.tsx
@@ -7,68 +7,19 @@ import type { ReactNode } from 'react';
 /**
  * WordPress dependencies
  */
-import { Spinner } from '@wordpress/components';
-import {
-	Component,
-	Suspense,
-	forwardRef,
-	useId,
-	useMemo,
-} from '@wordpress/element';
+import { forwardRef, useId, useMemo } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
 import { plugins } from '@wordpress/icons';
-// eslint-disable-next-line @wordpress/use-recommended-components
-import { Card, Icon, Stack, Notice, Text, VisuallyHidden } from '@wordpress/ui';
-import type { WidgetType } from '@wordpress/widget-primitives';
+import { Card, Icon, Stack, Text } from '@wordpress/ui';
 
 /**
  * Internal dependencies
  */
 import { useDashboardInternalContext } from '../../context/dashboard-context';
 import { WidgetContextProvider } from '../../context/widget-context';
-import { WidgetRender } from '../widget-render';
+import { LoadingOverlay, WidgetFrame } from '../widget-frame';
 import styles from './widget-chrome.module.css';
 import type { DashboardWidget } from '../../types';
-
-interface ErrorBoundaryProps {
-	children: ReactNode;
-}
-
-interface ErrorBoundaryState {
-	hasError: boolean;
-}
-
-class WidgetErrorBoundary extends Component<
-	ErrorBoundaryProps,
-	ErrorBoundaryState
-> {
-	state: ErrorBoundaryState = { hasError: false };
-
-	static getDerivedStateFromError(): ErrorBoundaryState {
-		return { hasError: true };
-	}
-
-	render() {
-		if ( this.state.hasError ) {
-			return (
-				<Notice.Root intent="error">
-					<Notice.Description>
-						{ __( 'This widget encountered an error.' ) }
-					</Notice.Description>
-				</Notice.Root>
-			);
-		}
-		return this.props.children;
-	}
-}
-
-function LoadingOverlay() {
-	return (
-		<Stack justify="center" align="center" className={ styles.loading }>
-			<Spinner />
-		</Stack>
-	);
-}
 
 interface UnavailableWidgetProps {
 	widgetTypeName: string;
@@ -101,53 +52,20 @@ function UnavailableWidget( { widgetTypeName }: UnavailableWidgetProps ) {
 	);
 }
 
-interface HeaderProps {
-	titleId: string;
-	widgetType: WidgetType;
-}
-
-function Header( { titleId, widgetType }: HeaderProps ) {
-	if ( ! widgetType.title ) {
-		return null;
-	}
-
-	return (
-		<Card.Header>
-			<Stack direction="row" align="center" gap="sm">
-				{ widgetType.icon && (
-					<span
-						className={ styles.widgetChromeHeaderIcon }
-						aria-hidden="true"
-					>
-						<Icon icon={ widgetType.icon } />
-					</span>
-				) }
-				<Card.Title id={ titleId } render={ <h2 /> }>
-					{ widgetType.title }
-				</Card.Title>
-			</Stack>
-		</Card.Header>
-	);
-}
-
 export interface WidgetChromeProps {
 	widget: DashboardWidget< unknown >;
 	index: number;
 	/**
-	 * Lifted by the surrounding `@wordpress/grid` surface into a sibling
-	 * slot of the grid item; not rendered by `WidgetChrome` itself.
-	 * Living outside `Card.Root` is what keeps these controls interactive
-	 * while edit mode applies `inert` to the chrome.
+	 * Lifted by `@wordpress/grid` into a sibling slot outside `Card.Root`, so
+	 * it stays interactive while edit mode makes the chrome `inert`.
 	 */
 	actionableArea?: ReactNode;
 	className?: string;
 }
 
 /**
- * Per-instance wrapper. Owns the chrome around a widget instance: identity
- * context, header (title + icon), edit-mode `inert` attribute, and the
- * error/loading boundaries that keep neighbours mounted when one widget fails
- * or is still resolving.
+ * Dashboard chrome: wraps `WidgetFrame` in the grid tile's `Card.Root`, owning
+ * identity, edit-mode `inert`, and the missing/resolving states.
  */
 export const WidgetChrome = forwardRef< HTMLDivElement, WidgetChromeProps >(
 	function WidgetChrome( { widget, index, className }, ref ) {
@@ -200,23 +118,6 @@ export const WidgetChrome = forwardRef< HTMLDivElement, WidgetChromeProps >(
 			);
 		}
 
-		// `presentation` encodes two independent axes. `full-bleed` hides
-		// the header; both `full-bleed` and `content-bleed` let the body
-		// break out of the content padding.
-		const { presentation } = widgetType;
-		const isHeaderHidden = presentation === 'full-bleed';
-		const isBodyBleeding =
-			presentation === 'full-bleed' || presentation === 'content-bleed';
-		const header = <Header titleId={ titleId } widgetType={ widgetType } />;
-
-		const body = (
-			<WidgetErrorBoundary>
-				<Suspense fallback={ <LoadingOverlay /> }>
-					<WidgetRender widget={ widget } widgetType={ widgetType } />
-				</Suspense>
-			</WidgetErrorBoundary>
-		);
-
 		return (
 			<WidgetContextProvider value={ contextValue }>
 				<Card.Root
@@ -226,28 +127,11 @@ export const WidgetChrome = forwardRef< HTMLDivElement, WidgetChromeProps >(
 					aria-labelledby={ widgetType.title ? titleId : undefined }
 					{ ...( editMode ? { inert: 'true' } : {} ) }
 				>
-					{ isHeaderHidden ? (
-						<VisuallyHidden>{ header }</VisuallyHidden>
-					) : (
-						header
-					) }
-
-					<Card.Content
-						className={ clsx(
-							styles.widgetChromeContent,
-							isBodyBleeding && styles.widgetChromeContentBleed
-						) }
-					>
-						{ isBodyBleeding ? (
-							<Card.FullBleed
-								className={ styles.widgetChromeBleedScroll }
-							>
-								{ body }
-							</Card.FullBleed>
-						) : (
-							body
-						) }
-					</Card.Content>
+					<WidgetFrame
+						widget={ widget }
+						widgetType={ widgetType }
+						titleId={ titleId }
+					/>
 				</Card.Root>
 			</WidgetContextProvider>
 		);
