+++ packages/widget-dashboard/src/components/widget-frame/index.ts
@@ -1,0 +1,2 @@
+export { WidgetFrame, LoadingOverlay } from './widget-frame';
+export type { WidgetFrameProps } from './widget-frame';
