+++ packages/widget-dashboard/src/components/widget-frame/widget-frame.module.css
@@ -1,0 +1,29 @@
+.headerIcon {
+	display: inline-flex;
+	color: var(--wpds-color-foreground-content-neutral);
+}
+
+/* Scroll container that keeps the header pinned; min:0 bounds tall content. */
+.content {
+	flex: 1;
+	height: 100%;
+	overflow-y: auto;
+	min-width: 0;
+	min-height: 0;
+}
+
+/* Zero block padding so a bleeding body fills the card top to bottom. */
+.contentBleed {
+	padding-block: 0;
+}
+
+/* Re-establish the scroll container at the card edges; cancel Card.FullBleed's bottom margin. */
+.bleedScroll {
+	height: 100%;
+	overflow-y: auto;
+	margin-block-end: 0;
+}
+
+.loading {
+	height: 100%;
+}
